requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow == 9.1.0
Shapely == 1.8.5.post1
matplotlib == 3.5.2
neat == 0.4.1
numpy == 1.22.3
pyglet == 2.0.0
torch == 1.13.0
